Work in Progress moving Tx generator from fixed-point-gentx to master. Created a new branch from master. Cherry picked the following files from fixed-point-gentx   1) chain-and-ledger/shelley-spec-ledger-test/src/Test/Shelley/Spec/Ledger/BenchmarkFunctions.hs   2) hain-and-ledger/shelley-spec-ledger-test/bench/Shelley/Spec/Ledger/Bench/Gen.hs   3) chain-and-ledger/shelley-spec-ledger-test/bench/Main.hs   4) chain-and-ledger/shelley-spec-ledger-test/bench/BenchValidation.hs   5) chain-and-ledger/shelley-spec-ledger-test/src/Test/Shelley/Spec/Ledger/Generator/Utxo.hs began fixing all the problems. Made some changes to the Val.hs.

+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/EpochBoundary.hs
@@ -109,8 +109,8 @@ obligation ::
   Map (KeyHash 'StakePool era) (PoolParams era) ->
   Coin
 obligation pp rewards stakePools =
-  Val.scale (length rewards) (_keyDeposit pp)
-    <> Val.scale (length stakePools) (_poolDeposit pp)
+  Val.scale (fromIntegral (length rewards)) (_keyDeposit pp)
+    <> Val.scale (fromIntegral (length stakePools)) (_poolDeposit pp)
 
 -- | Calculate maximal pool reward
 maxPool :: PParams era -> Coin -> Rational -> Rational -> Coin
